+++ engines/ep/tests/module_tests/collections/evp_store_collections_eraser_test.cc
@@ -1345,7 +1345,7 @@ protected:
             return vbstate->onDiskPrepares;
         };
 
-        int initPendingDocOnDisk = 0;
+        size_t initPendingDocOnDisk = 0;
         if (persistent() && !isMagma()) {
             initPendingDocOnDisk = getOnDiskPrepares();
         }
